lib modifications
