+++ lib/img_lib.py
